Adding insert persisting in an index table

+++ src/main/java/com/wlu/orm/hbase/schema/value/DoubleValue.java
@@ -13,13 +13,13 @@ public class DoubleValue implements Value {
 
 	
 
-	public double getIntValue() {
+	public double getDoubleValue() {
 		return doubleValue;
 	}
 
 
 
-	public void setIntValue(double intValue) {
+	public void setDoubleValue(double intValue) {
 		this.doubleValue = intValue;
 	}
